% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcount.R
\docType{package}
\name{gcount}
\alias{gcount}
\alias{gcount-package}
\title{gcount: Obtain Raw Read Counts from RNASeq Data}
\description{
\code{gcount} is an R-package that allows to obtain \emph{raw} read counts 
quickly and easily from RNASeq data to be used in downstream analyses.
}
\details{
The current functionality is as follows:

\itemize{
\item Provide a \code{bam/bed} file or object along with gtf/gff 
file or object as input.
\item Specify sequencing type: \code{paired} or \code{single} end data.
\item Specify library type: \code{unstranded}, \code{first-strand} or 
\code{second-strand} specific.
\item Filter reads based on number of mismatches prior to counting if 
necessary.
\item Consider or ignore reads that map to \emph{overlapping genes}.
\item Count reads that overlap feature of interest -- \code{genes}, 
\code{exons}, \code{introns} etc.
}
}
\seealso{
\code{\link{get_counts}}
}

