% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-counts.R
\name{get_counts}
\alias{get_counts}
\title{Obtain read counts from \code{bam} object.}
\usage{
get_counts(reads, annotation, transcript_id = "transcript_id",
  gene_id = "gene_id", mismatches = -1L, minoverlap = 1L,
  feature = c("gene_exon", "gene", "exon", "intron"), type = c("default",
  "union", "disjoin", "intersect", "longest", "shortest", "overlap"),
  library = c("unstranded", "firststrand", "secondstrand"), paired = FALSE,
  multiple_feature_overlaps = FALSE, verbose = FALSE)
}
\arguments{
\item{reads}{A \code{bam/bed} object (see \code{gread::read_format}) or 
complete path to a \code{bam/bed} file.}

\item{annotation}{A \code{gtf/gff} object (see \code{gread::read_format}) 
or complete path to \code{gtf/gff} file.}

\item{transcript_id}{Column name in \code{x} corresponding to transcript 
id. Default value is \code{"transcript_id"}.}

\item{gene_id}{Column name in \code{x} corresponding to gene id. Default 
value is \code{"gene_id"}.}

\item{mismatches}{Default \code{-1} is to ignore mismatches. If \code{> 0}, 
reads with mismatches \code{<= mismatches} are alone retained.}

\item{minoverlap}{Argument that's passed to 
\code{GenomicRanges::findOverlaps}.
Default is \code{1L}, i.e., select all overlapping reads.}

\item{feature}{How to count reads?

\code{"gene_exon"} counts only exonic reads within genes.

\code{"gene"} counts any/all reads overlapping the gene.

\code{"exon"} returns read counts for each exon separately.

\code{"intron"} returns read counts for each intron separately.

See \code{type} argument for more advanced operations on extracting 
feature coordinates.}

\item{type}{Same as \code{gread::extract}. See \code{?extract} in 
\code{gread}.}

\item{library}{Either \code{"unstranded"} (default), \code{"firststrand"} 
or \code{"secondstrand"}.}

\item{paired}{Default is \code{FALSE}. If the library is \code{paired end}, 
set it to \code{TRUE}}

\item{multiple_feature_overlaps}{logical. Should reads that overlap 
multiple features be counted. Default is \code{FALSE}, i.e., to discard. If 
\code{TRUE}, reads across overlapping features will be counted.}

\item{verbose}{logical. Default is \code{FALSE}. If \code{TRUE}, provides 
helpful messages to the console.}
}
\value{
A data.table with calculated raw counts of overlapping reads 
for each feature.
}
\description{
\code{get_counts} allows to obtain read counts quickly and 
easily from RNASeq data to be used in downstream analyses.
}
\examples{
path = system.file("tests", package="gcount")
gtf_file = file.path(path, "sample.gtf")
bam_file = file.path(path, "sample.bam")
bam_counts = get_counts(bam_file, gtf_file, feature="gene_exon", 
             type="union", paired=FALSE, library="unstranded", 
             verbose=TRUE)
}

