% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-gcount.R
\name{test_gcount}
\alias{test_gcount}
\title{Run a set of tests}
\usage{
test_gcount(verbose = FALSE, pkg = "pkg", silent = FALSE)
}
\arguments{
\item{verbose}{If \code{TRUE} sets \code{gcount_verbose} to \code{TRUE} for 
the duration of the tests.}

\item{pkg}{Root directory name under which all package content 
(ex: \code{DESCRIPTION}, \code{src/}, \code{R/}, \code{inst/} etc..) 
resides.}

\item{silent}{Logical, default \code{FALSE}. When \code{TRUE} it will not 
raise error on in case of test fails.}
}
\value{
When silent equals to \code{TRUE} it will return \code{TRUE} if 
all tests were successful. \code{FALSE} otherwise. If silent equals to 
\code{FALSE} it will return \code{TRUE} if all tests were successful. Error 
otherwise.
}
\description{
Runs a set of tests to check if \code{gcount} is working 
correctly.
}
\details{
Runs a series of tests. These can be used to see features and 
examples of usage, too. Running \code{test_gcount} will tell you the full 
location of the test file(s) to open.
}
\examples{
\dontrun{
gcount:::test_gcount()
}
}
\seealso{
\code{\link{get_counts}}
}

